requirements.txt
llama-cpp-python
openai-whisper
pyaudio
pyautogui
pynput
pyobjc-framework-ApplicationServices
pyobjc-framework-Quartz
pyperclip
python-dotenv
# These dependencies are not strictly required but enhance functionality
# For speech synthesis and assistant features
pyttsx3
requests
